refactor: rvr 64bit address ptr (#2928)
refactor address pointer related variables to 64bit instead of 32bit.
this PR modifies multiple files:
1) rvr core c and trace headers
2) ffi files
3) rvr code for extensions


reth benchmark result (original 32bit address space variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28121173062
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122726878

reth benchmark result (64bit address pointer variables):
1) https://github.com/openvm-org/rvr-openvm/actions/runs/28119229848
2) https://github.com/openvm-org/rvr-openvm/actions/runs/28122157089

By comparing "Exec Median (ms)" columns, it seems that 64bit version is
`10ms` faster or slower compared with the original 32bit version (in
both pure and metered modes). I am not sure why, it could be system
noises affecting it. Also, the comparison only compares execution speed,
excluding other possible important metrics when evaluating this pr.

resolves INT-8321

---------

+++ crates/vm/src/arch/rvr/io.rs
@@ -32,7 +32,7 @@ pub struct OpenVmIoState<'a, F: PrimeField32> {
 /// matching the C-side `abort_oob` termination used by `rd_mem_*`/`wr_mem_*`.
 /// Compiles to a no-op under the `unprotected` feature.
 #[cfg(not(feature = "unprotected"))]
-pub fn check_mem_bounds_range(start: u32, num_bytes: usize) {
+pub fn check_mem_bounds_range(start: u64, num_bytes: usize) {
     let start = start as usize;
     if start > MEM_SIZE || num_bytes > MEM_SIZE - start {
         panic!(
@@ -43,7 +43,7 @@ pub fn check_mem_bounds_range(start: u32, num_bytes: usize) {
 
 #[cfg(feature = "unprotected")]
 #[inline(always)]
-pub fn check_mem_bounds_range(_start: u32, _num_bytes: usize) {}
+pub fn check_mem_bounds_range(_start: u64, _num_bytes: usize) {}
 
 /// Replace the hint stream contents. Called via `ext_hint_stream_set` from extension FFI.
 ///
