+++ extensions/keccak256/rvr/c/rvr_ext_keccak.h
@@ -11,12 +11,12 @@ struct RvState;
  * caller convention. */
 
 extern __attribute__((preserve_most)) void rvr_ext_keccakf(
-    RvState* state, uint32_t buffer_ptr, uint32_t op_chip_idx,
+    RvState* state, uint64_t buffer_ptr, uint32_t op_chip_idx,
     uint32_t perm_chip_idx);
 
 extern __attribute__((preserve_most)) void rvr_ext_xorin(RvState* state,
-                                                         uint32_t buffer_ptr,
-                                                         uint32_t input_ptr,
+                                                         uint64_t buffer_ptr,
+                                                         uint64_t input_ptr,
                                                          uint32_t len,
                                                          uint32_t chip_idx);
